+++ changelog.md
@@ -57,6 +57,7 @@ Codion Change Log
 - EntityTableExportModel.populate() bug fixed, now handles missing attributes.
 - EntityTableExportModel.ExportTreeModel.showHidden() added, related refactoring.
 - ExportTreeModel refactored from EntityTableExportModel and renamed EntityTableExportTreeModel.
+- EntityTableExportTreeModel, showHidden no longer resets tree model.
 
 ## 0.18.62
 ### is.codion.common.reactive